+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [tbd]
+### Added
+- Screenshots: `screenshot/{2,3}` function and `screenshot: true` config for auto-capture
+
 ## [0.3.0] 2025-01-26
 ### Changed
 - Auto-convert case of playwright messages keys (snake_case to camelCase)